edit the folder structure in favlist
