+++ src/app/(events)/components/EventListCard/index.tsx
@@ -8,23 +8,61 @@ import { useSession } from "next-auth/react";
 import { IconHeart } from "@tabler/icons-react";
 import { BiTrash } from "react-icons/bi";
 import Link from "next/link";
+import { useCallback } from "react";
+import { trpc } from "~/providers";
+import { convertLocaleTimeString } from "~/utils/date-converter";
+import type { EventList } from "~/types";
 
 type CardNameType = "ADD" | "DELETE";
 
 interface EventListCardProps {
   event: ModifiedResult;
   cardName: CardNameType;
-  onClick: (event: ModifiedResult) => void;
+  setEventAdded?: (b: boolean) => void;
 }
 
 const EventListCard: React.FC<EventListCardProps> = ({
   event,
-  onClick,
   cardName,
+  setEventAdded,
 }) => {
   const { classes } = useStyles();
   const theme = useMantineTheme();
+
   const { data: session } = useSession();
+
+  const { mutate: mutateAdd } = trpc.favoriteEvents.addFavorite.useMutation();
+
+  const { mutate: mutateDelete } =
+    trpc.favoriteEvents.deleteFavorite.useMutation();
+
+  const handleAddFavEvent: (event: EventList) => void = useCallback(
+    (event) => {
+      const convertedDate = convertLocaleTimeString(event.dateTime);
+      setEventAdded && setEventAdded(true);
+      mutateAdd({
+        id: event.id,
+        userId: session?.user.id as string,
+        date: convertedDate,
+      });
+    },
+    [setEventAdded, mutateAdd, session?.user.id]
+  );
+
+  const handleDeleteFavEvent: (event: ModifiedResult) => void = useCallback(
+    (event) => {
+      mutateDelete({
+        id: event.id,
+      });
+    },
+    [mutateDelete]
+  );
+
+  const onClickIcon = (e: React.MouseEvent<HTMLButtonElement>) => {
+    e.stopPropagation();
+    e.preventDefault();
+    cardName === "ADD" ? handleAddFavEvent(event) : handleDeleteFavEvent(event);
+  };
 
   return (
     <Link
@@ -71,10 +109,10 @@ const EventListCard: React.FC<EventListCardProps> = ({
 
             {session && (
               <ActionIcon
-                onClick={() => onClick(event)}
                 className={classes.icon}
+                onClick={(e) => onClickIcon(e)}
               >
-                {cardName == "ADD" ? (
+                {cardName === "ADD" ? (
                   <IconHeart fill="#ee6c4d" stroke="#ee6c4d" />
                 ) : (
                   <BiTrash fill="#555" />
