+++ src/app/(events)/favlist/page.tsx
@@ -1,0 +1,10 @@
+import { getServerAuthSession } from "~/server/auth";
+import ClientFavListPage from "./components/ClientFavListPage";
+
+const FavListPage = async () => {
+  const session = await getServerAuthSession();
+
+  return <ClientFavListPage session={session} />;
+};
+
+export default FavListPage;
