+++ src/types/index.d.ts
@@ -6,3 +6,20 @@ export interface Event {
   imageUrl: string;
   website: string;
 }
+
+export interface EventList {
+  id: string;
+  title: string;
+  eventUrl: string;
+  description: string;
+  venue?: {
+    id?: string;
+    name?: string;
+    address?: string;
+    lat?: number;
+    lng?: number;
+  } | null;
+  imageUrl: string;
+  imageId: string;
+  dateTime: string;
+}
